PTAS Change requests
1. A prompt needs to be added for when the user tries to skip responding
to an item. It should follow the same rules that are used in Abstract
Reasoning, Number Series, and Visualization in 2D.

2. A prompt needs to be added for when the user does not submit a
response within 120 sec (same rules that are used in AR, NS, and V2D).

3. Simple Reaction Time test – please add 5 practice trials at the
beginning of the test. Collect the same data for practice trials that is
collected for test trials.

4. Spatial Visualization – the images for the P and L sample items are
fuzzy.

5. Misc. – the pop-up that appears at the end of each measure is really
small. Make larger font and next button.

6. For Abstract Reasoning, Number Series, and Spatial Visualization, the
“Next” button should be inactive for the first 5 seconds on all test
items. This is to prevent the user from just clicking through the test
as fast as possible.